Refactor network representation
Originally cider was extracted from some code I wrote to solve a
programming interview question we have at work, so it was uniquely
tailored to that specific problem. Then I created the `cider` tool, and
it became more generally useful. However, the library API was still a
bit of a mess, and some operations, such as concatenating two CIDR
blocks, didn't really make sense in a general setting.

Thus the Data.IP.IPv4 module has been completely refactored to better
and more clearly represent networks, and expose the properties and
operations that are generally useful in networks. It will allow me to
rework the `cider` tool itself to present more of the information
described by the tools in [1].

  [1] https://www.cyberciti.biz/faq/linux-subnet-calculator-cidr/